Node class update and Gaussian functions
Updated the perpendicular optimization function in the Node class. It now
completes without errors, but there may still be problems with the scaling
factor becoming unstable. Added gaussian.py, which contains the functions
for reading energies and forces from Gaussian logfiles. test.py contains
short code that can be used to test the optimization in the Node class.