Removal of parallel LST path generation.
The generation of the LST path in parallel resulted in severe overloading
of compute nodes, likely because bound methods were being parallelized in
a non-standard way. This functionality has been removed, which will
result in slightly slower code, but should not matter much since the main
computational expense are the quantum calculations.