Changed Node class.
`getBonds` was added to the Node class, which enables reading the bonds of
the structure directly from the coordinates without first going through a
`gen3D.Molecule` object. A new class, Atom, was implemented to enable
this. Additionally, indexing of the `gen3D.Molecule` class was changed to
start indexing from 0.